Moved helpers to ModLoader.Utilities namespace.
