+++ ExampleMod/Content/NPCs/PartyZombie.cs
@@ -4,6 +4,7 @@ using Terraria.ModLoader;
 using Terraria.GameContent.Bestiary;
 using Terraria.GameContent.ItemDropRules;
 using ExampleMod.Content.Biomes;
+using Terraria.ModLoader.Utilities;
 
 namespace ExampleMod.Content.NPCs
 {
